add authority in user
