+++ src/app/AboutMe/page.tsx
@@ -1,16 +1,12 @@
-import { INDIGO, INFO } from "@/lib/constants";
+import { ADMIN, INDIGO } from "@/lib/constants";
 import { MdPerson } from "react-icons/md";
 import UserAvatar from "../_components/UserAvatar";
 import { auth } from "@/lib/auth";
-import { Button } from "@/components/ui/button";
-import { Label } from "@/components/ui/label";
-import LabelText from "./_components/LabelText";
 import ChangePasswordDialog from "./_components/ChangePasswordDialog";
+import AboutAdminInfo from "./_components/AboutAdminInfo";
+import AboutMeInfo from "./_components/AboutMeInfo";
 export default async function AboutMe(){
   const session = await auth();
-  const infoLabels = ["學校/單位","學校職務","科目/課程","聯盟角色"];
-  const index = INFO.findIndex(({ email }) => email === session?.user?.email);
-  const info = INFO[index].info;
   return(
     <div className="grid h-full p-16">
       <a className="mt-4 mb-8 inline-flex items-center text-sm font-medium font-semibold gap-1 text-gray-700 hover:opacity-80">
@@ -28,46 +24,9 @@ export default async function AboutMe(){
           >
             登出
           </a>
+          { session?.user?.authority === ADMIN && <ChangePasswordDialog /> }
         </div>
-        <div className="w-3/4 flex flex-col gap-3 items-center">
-          <p className="text-2xl" style={{color: INDIGO}}>個人資料</p>
-          <div className="rounded-3xl border-2 w-full px-4 py-6 gap-3 flex flex-col" style={{borderColor: INDIGO}}>
-            {infoLabels.map((label,i) => (
-              <LabelText label={label} text={info[i]} />
-            ))}
-            <div className="flex items-center">
-              <a className="font-semibold mr-1">教學紀錄：</a>
-              <a 
-                className="justify-self-center w-fit px-3 py-1 text-white text-sm rounded-full hover:opacity-80"
-                style={{backgroundColor: INDIGO}}
-              >
-                查看教學年資表
-              </a>
-            </div>
-          </div>
-          <p className="pt-3 text-2xl" style={{color: INDIGO}}>帳號安全性與隱私管理</p>
-          <div className="rounded-3xl border-2 w-full p-4 flex-col flex gap-3" style={{borderColor: INDIGO}}>
-            <div className="grid grid-cols-3">
-              <LabelText label={"帳號"} text={session?.user?.email ?? ""} />
-              <LabelText label={"手機號碼"} text={session?.user?.mobile ?? ""} />
-            </div>
-            <div className="flex gap-3">
-              {/* <a 
-                className="justify-self-center w-fit px-3 py-1 text-white text-sm rounded-full hover:opacity-80"
-                style={{backgroundColor: INDIGO}}
-              >
-                變更密碼
-              </a> */}
-              <ChangePasswordDialog />
-              <a 
-                className="justify-self-center w-fit px-3 py-1 text-white text-sm rounded-full hover:opacity-80"
-                style={{backgroundColor: INDIGO}}
-              >
-                查看隱私管理辦法
-              </a>
-            </div>
-          </div>
-        </div>
+        { session?.user?.authority === ADMIN ? <AboutAdminInfo /> : <AboutMeInfo /> }
       </div>
     </div>
   );
