+++ src/lib/auth/index.ts
@@ -8,27 +8,16 @@ export const {
 } = NextAuth({
   providers: [CredentialsProvider],
   callbacks: {
-    // async jwt({ token, user }) {
-    //   if (user) {
-    //     token.name = user.username;
-    //     token.email = user.email;
-    //     token.mobile = user.mobile;
-    //   }
-    //   return token;
-    // },
     async session({ session, token }) {
       const userEmail = token.email || session?.user?.email;
       if (!userEmail) return session;
-      // const email = "myemail@gmail.com";
-      // const username = "阿依";
-      // const mobile = "0912345678";
       const user = USERS[USERS.findIndex(({ email }) => email === userEmail)];
       if (session.user) {
         session.user.username = user.username;
         session.user.email = user.email;
         session.user.mobile = user.mobile;
+        session.user.authority = user.authority;
       }
-      console.log(session)
       return session;
     },
   },
